+++ src/eca/config.clj
@@ -219,6 +219,7 @@
                          :deny {}}
               :readFile {:maxLines 2000}
               :shellCommand {:summaryMaxLength 35}
+              :editorNav {:enabled true}
               :outputTruncation {:lines 2000 :sizeKb 50}}
    :variantsByModel {".*sonnet[-._]4[-._]6|opus[-._]4[-._][56]" {:variants anthropic-variants
                                                                  :api ["anthropic" "bedrock"]}